fix soliaaris compiling error

+++ src/CAR-ALASSO.cpp
@@ -68,14 +68,14 @@ List CAR_ALASSO_Cpp(const arma::mat & data, // col composition data, ROW as a sa
   int p = design.n_cols; //number of predictors
   int n = data.n_rows; // number of samples
   
-  int n_save = floor(n_iter/thin_by); //
+  int n_save = floor(static_cast<double>(n_iter/thin_by)); //
   int i_save = 0;  
   
   // mcmc matrices:
   arma::mat beta_mcmc(n_save,k * p); // beta mcmc
   beta_mcmc += NA_REAL; 
   
-  arma::mat Omega_mcmc(n_save , floor( k * (k+1)/2 )  ); // vectorized column first, but had no diagnol
+  arma::mat Omega_mcmc(n_save , floor( static_cast<double>(k * (k+1)/2) )  ); // vectorized column first, but had no diagnol
   Omega_mcmc += NA_REAL;
